+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - Editors and Viewers can now be invited to projects
   - Randomizations now keep track of the user who randomized the subject
   - Emails are sent when subjects are randomized
+  - Randomizations now show the randomization number of the total randomization goal
 
 ## 0.0.0 (August 21, 2013)